--- module #3. REST API Advanced/core/src/main/java/com/epam/esm/config/SpringConfig.java
@@ -1,78 +1,0 @@
-package com.epam.esm.config;
-
-import org.apache.commons.dbcp2.BasicDataSource;
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.context.support.ResourceBundleMessageSource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
-import org.springframework.validation.beanvalidation.MethodValidationPostProcessor;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.i18n.AcceptHeaderLocaleResolver;
-import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
-
-import javax.validation.Validator;
-import java.util.Locale;
-
-@Configuration
-@EnableWebMvc
-@ComponentScan("com.epam.esm")
-public class SpringConfig{
-
-    @Bean
-    public JdbcTemplate jdbcTemplate(BasicDataSource dataSource){
-        return new JdbcTemplate(dataSource);
-    }
-
-    @Bean
-    public BasicDataSource dataSource(DbProperties dbProperties) {
-        BasicDataSource ds = new BasicDataSource();
-        ds.setDriverClassName(dbProperties.getDriver());
-        ds.setUrl(dbProperties.getUrl());
-        ds.setUsername(dbProperties.getUser());
-        ds.setPassword(dbProperties.getPassword());
-        return ds;
-    }
-
-    @Bean
-    private static PropertySourcesPlaceholderConfigurer propertyConfigInDev() {
-        return new PropertySourcesPlaceholderConfigurer();
-    }
-
-    @Bean
-    public MethodValidationPostProcessor validationPostProcessor(Validator validator) {
-        MethodValidationPostProcessor postProcessor = new MethodValidationPostProcessor();
-        postProcessor.setValidator(validator);
-        return postProcessor;
-    }
-
-    @Bean
-    public LocalValidatorFactoryBean validator() {
-        return new LocalValidatorFactoryBean();
-    }
-
-    @Bean
-    public MessageSource messageSource() {
-        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
-        messageSource.setBasenames("lang/message");
-        messageSource.setDefaultEncoding("UTF-8");
-        return messageSource;
-    }
-
-    @Bean
-    public AcceptHeaderLocaleResolver headerResolver(){
-        AcceptHeaderLocaleResolver headerResolver = new AcceptHeaderLocaleResolver();
-        headerResolver.setDefaultLocale(Locale.ENGLISH);
-        return headerResolver;
-    }
-
-    @Bean
-    public LocaleChangeInterceptor interceptor(){
-        LocaleChangeInterceptor interceptor = new LocaleChangeInterceptor();
-        interceptor.setParamName("lang");
-        return interceptor;
-    }
-}
