+++ module #3. REST API Advanced/core/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+spring.datasource.url=jdbc:mysql://localhost:3306/certificate?serverTimezone=Europe/Minsk
+spring.datasource.username=${DB_USER}
+spring.datasource.password=${DB_PASS}
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
