+++ module #3. REST API Advanced/web/src/main/resources/lang/message_ru.properties
